syntax

+++ adv_gen_own.py
@@ -74,8 +74,8 @@ def main(argv=None):
 
 
     # preprocess the data
-    # X_train = X_train.astype('float32')
-    # X_test = X_test.astype('float32')
+    X_train = X_train.astype('float32')
+    X_test = X_test.astype('float32')
     X_train /= 255.0
     X_test /= 255.0
